implement layout autotuning for amp training
